SettingManager UpdateCommand Client Done

+++ Manager/ManagerService/Console/ClientService.cs
@@ -21,26 +21,41 @@ namespace ManagerService.Console
             List<DataPermission> dataPermissions = new List<DataPermission>();
             try
             {
-                Guid userId = UserDataAccess.Login(userName, password,out dataPermissions);
-                if (userId != Guid.Empty)
+                User user = UserDataAccess.Login(userName, password,out dataPermissions);
+                if (user.UserId != Guid.Empty)
                 {
                     Dictionary<string, List<Guid>> accountPermissions = new Dictionary<string, List<Guid>>();
                     Dictionary<string, List<Guid>> instrumentPermissions = new Dictionary<string, List<Guid>>();
-                    foreach (DataPermission data in dataPermissions)
+                    foreach (ExchangeSystemSetting item in Manager.ManagerSettings.ExchangeSystems)
                     {
-                        List<Guid> memberIds = ExchangeData.GetMemberIds(data.ExchangeSystemCode, data.DataObjectId);
-                        if (data.DataObjectType ==  DataObjectType.Account)
+                        List<Guid> accountMemberIds = new List<Guid>();
+                        List<Guid> instrumentMemberIds = new List<Guid>();
+                        List<DataPermission> systemPermissions = dataPermissions.FindAll(delegate(DataPermission data)
                         {
-                            accountPermissions.Add(data.ExchangeSystemCode, memberIds);
-                        }
-                        else
+                            return data.ExchangeSystemCode == item.Code;
+                        });
+                        if (systemPermissions.Count>0)
                         {
-                            instrumentPermissions.Add(data.ExchangeSystemCode, memberIds);
+                            foreach (DataPermission permission in systemPermissions)
+                            {
+                                List<Guid> memberIds = new List<Guid>();
+                                memberIds.AddRange(ExchangeData.GetMemberIds(permission.ExchangeSystemCode, permission.DataObjectId));
+                                if (permission.DataObjectType == DataObjectType.Account)
+                                {
+                                    accountMemberIds.AddRange(memberIds);
+                                }
+                                else if(permission.DataObjectType == DataObjectType.Instrument)
+                                {
+                                    instrumentMemberIds.AddRange(memberIds);
+                                }
+                            }
+                            accountPermissions.Add(item.Code, accountMemberIds);
+                            instrumentPermissions.Add(item.Code, instrumentMemberIds);
                         }
                     }
                     string sessionId = OperationContext.Current.SessionId;
                     IClientProxy clientProxy = OperationContext.Current.GetCallbackChannel<IClientProxy>();
-                    this._Client = Manager.ClientManager.AddClient(oldSessionId, sessionId, userId, clientProxy, language,accountPermissions,instrumentPermissions);
+                    this._Client = Manager.ClientManager.AddClient(oldSessionId, sessionId, user, clientProxy, language,accountPermissions,instrumentPermissions);
 
                     OperationContext.Current.Channel.Faulted += this._Client.Channel_Broken;
                     OperationContext.Current.Channel.Closed += this._Client.Channel_Broken;
@@ -77,7 +92,7 @@ namespace ManagerService.Console
             }
             catch (Exception ex)
             {
-                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "GetFunctionTree.\r\n{0}", ex.ToString());
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/GetFunctionTree.\r\n{0}", ex.ToString());
                 return new FunctionTree();
             }
         }
@@ -92,27 +107,51 @@ namespace ManagerService.Console
             }
             catch (Exception ex)
             {
-                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ChangePassword.\r\n{0}", ex.ToString());
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/ChangePassword.\r\n{0}", ex.ToString());
                 return false;
             }
         }
 
         public List<AccessPermission> GetAccessPermissions()
         {
-            List<AccessPermission> permissions = new List<AccessPermission>();
-            return permissions;
+            try
+            {
+                List<AccessPermission> permissions = UserDataAccess.GetAccessPermissions(this._Client.userId, this._Client.language);
+                return permissions;
+            }
+            catch (Exception ex)
+            {
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/GetAccessPermission.\r\n{0}", ex.ToString());
+                return null;
+            }
         }
 
         public List<UserData> GetUserData()
         {
-            List<UserData> data = UserDataAccess.GetUserData();
-            return data;
+            try
+            {
+                List<UserData> data = UserDataAccess.GetUserData();
+                return data;
+            }
+            catch (Exception ex)
+            {
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/GetUserData.\r\n{0}", ex.ToString());
+                return null;
+            }
         }
 
         public List<RoleData> GetRoles()
         {
-            List<RoleData> roles = UserDataAccess.GetRoles(this._Client.language.ToString());
-            return roles;
+            try
+            {
+                List<RoleData> roles = UserDataAccess.GetRoles(this._Client.language.ToString());
+                return roles;
+            }
+            catch (Exception ex)
+            {
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/GetRoles.\r\n{0}", ex.ToString());
+                return null;
+            }
         }
 
         public bool UpdateUsers(UserData user, string password, bool isNewUser)
@@ -132,7 +171,7 @@ namespace ManagerService.Console
             }
             catch (Exception ex)
             {
-                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "UpdateUsers.\r\n{0}", ex.ToString());
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/UpdateUsers.\r\n{0}", ex.ToString());
                 return false;
             }
         }
@@ -141,11 +180,18 @@ namespace ManagerService.Console
         {
             try
             {
-                return UserDataAccess.DeleteUser(userId);
+                if (userId == this._Client.userId)
+                {
+                    return false;
+                }
+                else
+                {
+                    return UserDataAccess.DeleteUser(userId);
+                }
             }
             catch (Exception ex)
             {
-                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ClientServer/DeleteUser.\r\n{0}", ex.ToString());
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/DeleteUser.\r\n{0}", ex.ToString());
                 return false;
             }
         }
@@ -159,8 +205,51 @@ namespace ManagerService.Console
             }
             catch (Exception ex)
             {
-                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "GetAllPermission.\r\n{0}", ex.ToString());
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/GetAllPermission.\r\n{0}", ex.ToString());
                 return null;
+            }
+        }
+
+        public bool UpdateRole(RoleData role, bool isNewRole)
+        {
+            try
+            {
+                bool isSuccess = false;
+                if (isNewRole)
+                {
+                    isSuccess = UserDataAccess.AddNewRole(role);
+                }
+                else
+                {
+                    isSuccess = UserDataAccess.EditRole(role);
+                }
+                return isSuccess;
+                
+            }
+            catch (Exception ex)
+            {
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/UpdateRole.\r\n{0}", ex.ToString());
+                return false;
+            }
+        }
+
+        public bool DeleteRole(int roleId)
+        {
+            try
+            {
+                if (this._Client.user.IsInRole(roleId))
+                {
+                    return false;
+                }
+                else
+                {
+                    return UserDataAccess.DeleteRole(roleId, this._Client.userId);
+                }
+            }
+            catch (Exception ex)
+            {
+                Logger.TraceEvent(System.Diagnostics.TraceEventType.Error, "ManagerService.Console.ClientService/DeleteRole.\r\n{0}", ex.ToString());
+                return false;
             }
         }
 
@@ -218,6 +307,22 @@ namespace ManagerService.Console
             catch (Exception ex)
             {
                 Logger.AddEvent(TraceEventType.Error, "ClientService.AcceptPlace error:\r\n{0}", ex.ToString());
+            }
+            return transactionError;
+        }
+
+        public TransactionError CancelPlace(Guid transactionId, CancelReason cancelReason)
+        {
+            TransactionError transactionError = TransactionError.OK;
+            string exchangeCode = "WF01";
+            try
+            {
+                ExchangeSystem exchangeSystem = Manager.ExchangeManager.GetExchangeSystem(exchangeCode);
+                transactionError = exchangeSystem.CancelPlace(transactionId,cancelReason);
+            }
+            catch (Exception ex)
+            {
+                Logger.AddEvent(TraceEventType.Error, "ClientService.CancelPlace error:\r\n{0}", ex.ToString());
             }
             return transactionError;
         }
